backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
joblib
pandas
scikit-learn
sqlalchemy
psycopg2-binary
python-dotenv
slowapi
